+++ alias.go
@@ -7,16 +7,7 @@ import (
 	"github.com/Chendemo12/flaskgo/internal/app"
 	"github.com/Chendemo12/flaskgo/internal/core"
 	"github.com/Chendemo12/flaskgo/internal/godantic"
-	"github.com/Chendemo12/flaskgo/internal/mode"
 	"time"
-)
-
-//goland:noinspection GoUnusedGlobalVariable
-const (
-	DevMode        = mode.DevMode
-	ProdMode       = mode.ProdMode
-	RouteSeparator = app.RouteSeparator
-	Version        = app.Version
 )
 
 //goland:noinspection GoUnusedGlobalVariable
@@ -28,45 +19,41 @@ var (
 	CombinePath       = app.CombinePath
 )
 
-//goland:noinspection GoUnusedGlobalVariable
-var (
-	IsDebug = mode.IsDebug
-	SetMode = mode.SetMode
-	GetMode = mode.GetMode
-)
-
 type Field = godantic.Field
 type BaseModel = godantic.BaseModel
 type BaseModelIface = godantic.Iface
 
 //goland:noinspection GoUnusedGlobalVariable
-//var ( // types
-//	Int8    = openapi.Int8
-//	Int16   = openapi.Int16
-//	Int32   = openapi.Int32
-//	Int64   = openapi.Int64
-//	Uint8   = openapi.Uint8
-//	Uint16  = openapi.Uint16
-//	Uint32  = openapi.Uint32
-//	Uint64  = openapi.Uint64
-//	Float32 = openapi.Float32
-//	Float64 = openapi.Float64
-//	String  = openapi.String
-//	Boolean = openapi.Boolean
-//	Bool    = openapi.Boolean
-//	Mapping = openapi.Mapping
-//
-//	Int     = Int32
-//	Byte    = Uint8
-//	Uint    = Uint32
-//	Float   = Float64
-//	Array   = openapi.List
-//	List    = openapi.List
-//	Ints    = &openapi.RouteModel{Model: Int32, Struct: Int32, RetArray: true}
-//	Bytes   = &openapi.RouteModel{Model: Uint8, Struct: Uint8, RetArray: true}
-//	Strings = &openapi.RouteModel{Model: String, Struct: String, RetArray: true}
-//	Floats  = &openapi.RouteModel{Model: Float64, Struct: Float64, RetArray: true}
-//)
+var ( // types
+	String = godantic.String
+	Str    = godantic.String
+
+	Bool    = godantic.Bool
+	Boolean = godantic.Bool
+
+	Int    = godantic.Int
+	Byte   = godantic.Uint8
+	Int8   = godantic.Int8
+	Int16  = godantic.Int16
+	Int32  = godantic.Int32
+	Int64  = godantic.Int64
+	Uint8  = godantic.Uint8
+	Uint16 = godantic.Uint16
+	Uint32 = godantic.Uint32
+	Uint64 = godantic.Uint64
+	// Float32 = openapi.Float32
+	// Float64 = openapi.Float64
+
+	// Mapping = openapi.Mapping
+
+	// Float   = Float64
+	// Array   = openapi.List
+	// List    = openapi.List
+	// Ints    = &openapi.RouteModel{Model: Int32, Struct: Int32, RetArray: true}
+	// Bytes   = &openapi.RouteModel{Model: Uint8, Struct: Uint8, RetArray: true}
+	// Strings = &openapi.RouteModel{Model: String, Struct: String, RetArray: true}
+	// Floats  = &openapi.RouteModel{Model: Float64, Struct: Float64, RetArray: true}
+)
 
 //goland:noinspection GoUnusedGlobalVariable
 type Dict = map[string]any // python dict
@@ -96,29 +83,24 @@ type Scheduler = app.Scheduler
 type CronJobFunc = app.CronJobFunc
 
 //goland:noinspection GoUnusedGlobalVariable
-//var (
-//	ValidationErrorResponse = app.ValidationErrorResponse
-//	AnyResponse             = app.AnyResponse
-//	JSONResponse            = app.JSONResponse
-//	StringResponse          = app.StringResponse
-//	StreamResponse          = app.StreamResponse
-//	FileResponse            = app.FileResponse
-//	ErrorResponse           = app.ErrorResponse
-//	HTMLResponse            = app.HTMLResponse
-//	OKResponse              = app.OKResponse
-//	ResourceNotFound        = app.ResourceNotFound
-//	AdvancedResponse        = app.AdvancedResponse
-//	StringsReverse          = openapi.StringsReverse
-//)
+var (
+	OKResponse              = app.OKResponse
+	JSONResponse            = app.JSONResponse
+	ErrorResponse           = app.ErrorResponse
+	ValidationErrorResponse = app.ValidationErrorResponse
+	AnyResponse             = app.AnyResponse
+	StringResponse          = app.StringResponse
+	StreamResponse          = app.StreamResponse
+	FileResponse            = app.FileResponse
+	HTMLResponse            = app.HTMLResponse
+	AdvancedResponse        = app.AdvancedResponse
+)
 
 // DisableBaseRoutes 禁用基础路由
 func DisableBaseRoutes() { core.BaseRoutesDisabled = true }
 
 // DisableSwagAutoCreate 禁用文档自动生成
 func DisableSwagAutoCreate() { core.SwaggerDisabled = true }
-
-// DisableDefaultOutput 禁用默认输出
-func DisableDefaultOutput() { core.PrintDisabled = true }
 
 // DisableRequestValidate 禁用请求体自动验证
 func DisableRequestValidate() { core.RequestValidateDisabled = true }